requirements.txt
Flask == 3.0.0 
mysql-connector-python == 8.1.0 
pandas == 2.1.1
openpyxl == 3.1.2 